Python Yolo/requirements.txt
ultralytics
opencv-python
pillow
requests
pyyaml
tqdm
pandas
matplotlib
seaborn